+++ Simulation/src/ste/engine/graphics_interface/command_buffer/command/cmd_copy_buffer.hpp
@@ -41,7 +41,7 @@ private:
 	void operator()(const command_buffer &command_buffer, command_recorder &) const override final {
 		vkCmdCopyBuffer(command_buffer, src_buffer.get().get_buffer_handle(),
 						dst_buffer.get().get_buffer_handle(),
-						ranges.size(), ranges.data());
+						static_cast<std::uint32_t>(ranges.size()), ranges.data());
 	}
 };
